+++ src/modules/tooltip/Position.js
@@ -183,10 +183,10 @@ export default class Position {
     const tooltipEl = ttCtx.getElTooltip()
     let tooltipRect = ttCtx.tooltipRect
 
-    let pointR = r !== null ? parseInt(r) : 1
+    let pointR = r !== null ? parseFloat(r) : 1
 
-    let x = parseInt(cx) + pointR + 5
-    let y = parseInt(cy) + pointR / 2 // - tooltipRect.ttHeight / 2
+    let x = parseFloat(cx) + pointR + 5
+    let y = parseFloat(cy) + pointR / 2 // - tooltipRect.ttHeight / 2
 
     if (x > w.globals.gridWidth / 2) {
       x = x - tooltipRect.ttWidth - pointR - 15
